+++ src/test/kotlin/com/mytaxi/apis/phraseapi/client/PhraseApiClientTest.kt
@@ -1,17 +1,17 @@
-package com.mytaxi.apis.phraseapi
+package com.mytaxi.apis.phraseapi.client
 
 import com.google.common.net.HttpHeaders
 import com.google.common.net.MediaType
 import com.google.gson.Gson
-import com.mytaxi.apis.phraseapi.locale.reponse.Message
-import com.mytaxi.apis.phraseapi.locale.reponse.PhraseLocale
-import com.mytaxi.apis.phraseapi.locale.reponse.PhraseLocaleMessages
-import com.mytaxi.apis.phraseapi.locale.reponse.PhraseLocales
-import com.mytaxi.apis.phraseapi.project.reponse.PhraseProject
+import com.mytaxi.apis.phraseapi.client.model.Message
+import com.mytaxi.apis.phraseapi.client.model.PhraseLocale
+import com.mytaxi.apis.phraseapi.client.model.PhraseLocaleMessages
+import com.mytaxi.apis.phraseapi.client.model.PhraseLocales
+import com.mytaxi.apis.phraseapi.client.model.PhraseProject
 import feign.Response
 import org.junit.Test
+import org.mockito.Mockito
 import org.mockito.Mockito.`when`
-import org.mockito.Mockito.mock
 import java.nio.charset.StandardCharsets
 import java.util.Locale
 import java.util.UUID
@@ -20,7 +20,7 @@ import kotlin.test.assertNotNull
 
 class PhraseApiClientTest {
 
-    private var client: PhraseApi = mock(PhraseApi::class.java)
+    private var client: PhraseApi = Mockito.mock(PhraseApi::class.java, Mockito.withSettings().extraInterfaces(CacheApi::class.java))
 
     private var phraseApiClient: PhraseApiClient
